+++ crates/star-bert/CHANGELOG.md
@@ -1,2 +1,3 @@
 # 0.1.0
 - Implement `integer`
+- Implement a custom error type for granularity.